requirements.txt
absl-py==2.3.0
aiohttp==3.9.3
aiosignal==1.3.2
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
asyncio==3.4.3
attrs==23.2.0
autobahn==19.11.2
Automat==22.10.0
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
boto3==1.38.46
botocore==1.38.46
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.4
colorama==0.4.6
comm==0.2.0
comtypes==1.4.11
constantly==23.10.4
construct==2.10.70
contourpy==1.2.0
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
empyrical==0.5.5
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==2.3.2
fonttools==4.46.0
fqdn==1.5.1
frozendict==2.3.10
frozenlist==1.7.0
fyers-api==1.0.9
fyers_apiv3==3.1.7
h11==0.14.0
h5py==3.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.10.0
kiteconnect==5.0.1
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mplfinance==0.12.10b0
multidict==6.6.2
multitasking==0.0.11
namex==0.1.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.2
openpyxl==3.1.5
optree==0.16.0
overrides==7.7.0
packaging==23.2
panda==0.3.1
pandas==2.1.4
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.3
peewee==3.17.0
Pillow==10.1.0
platformdirs==4.1.0
playsound==1.2.2
plotly==5.18.0
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.41
propcache==0.3.2
protobuf==5.29.3
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pycryptodomex==3.20.0
Pygments==2.17.2
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.1.1
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
PythonTurtle==0.3.2
pyttsx3==2.98
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.17.1
s3transfer==0.13.0
scikit-learn==1.5.2
scipy==1.12.0
Send2Trash==1.8.2
service-identity==24.2.0
six==1.16.0
smartapi-python==1.5.5
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
ta==0.11.0
tenacity==8.2.3
tensor==0.3.6
terminado==0.18.0
threadpoolctl==3.5.0
timestamp==0.0.1
tinycss2==1.2.1
tornado==6.1
traitlets==5.14.0
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==25.6.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.20
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==8.1
Werkzeug==2.3.6
wrapt==1.17.2
yarl==1.20.1
yfinance==0.2.33
zope.interface==6.1
